Cleaned up some data structures a bit more
git-svn-id: https://trac.beirdo.ca:4443/svn/havokmud/branches/rev_1_11@1615 69fa473d-82f3-0310-a870-f1ebe01cfa2b

+++ havokmud/include/externs.h
@@ -168,8 +168,8 @@ extern char    *drinknames[];
 extern int      qp_patience;
 
 extern int      auct_loop;
-extern long     minbid;
-extern long     intbid;
+extern int      minbid;
+extern int      intbid;
 extern struct char_data *auctioneer;
 extern struct char_data *bidder;
